+++ Applications/VRP/src/CutGen/shrink.c
@@ -220,8 +220,9 @@ int greedy_shrinking1(network *n, double truck_cap, double etol,
 	       new_cut->type = (set_size < vertnum/2 ?
 				SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	       new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-				RHS(set_size, set_demand, truck_cap):
-				2*BINS(set_demand, truck_cap));
+				RHS((int) set_size, (int) set_demand,
+				    (int) truck_cap):
+				2*BINS((int) set_demand, (int) truck_cap));
 	       for (k = 0, cutpt = cut_list; k < num_cuts; k++,
 		       cutpt += new_cut->size)
 		  if (!memcmp(coef, cutpt, new_cut->size*sizeof(char)))
@@ -363,8 +364,9 @@ int greedy_shrinking6(network *n, double truck_cap, double etol,
 	       new_cut->type = (set_size < vertnum/2 ?
 				SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	       new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-				RHS(set_size, set_demand, truck_cap):
-				2*BINS(set_demand, truck_cap));
+				RHS((int) set_size, (int) set_demand,
+				    (int) truck_cap):
+				2*BINS((int) set_demand, (int) truck_cap));
 	       
 	       for (k = 0, cutpt = cut_list; k < num_cuts; k++,
 		       cutpt += new_cut->size)
@@ -479,8 +481,9 @@ int greedy_shrinking1_one(network *n, double truck_cap, double etol,
 	    new_cut->type = (set_size < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(set_size,set_demand,truck_cap):
-			     2*BINS(set_demand,truck_cap));
+			     RHS((int) set_size, (int) set_demand,
+				 (int) truck_cap):
+			     2*BINS((int) set_demand, (int) truck_cap));
 	    for (k = 0, cutpt = cut_list; k < num_cuts; k++,
 		    cutpt += new_cut->size)
 		  if (!memcmp(coef, cutpt, new_cut->size*sizeof(char)))
@@ -517,8 +520,9 @@ int greedy_shrinking1_one(network *n, double truck_cap, double etol,
 	    new_cut->type = (complement_size < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE : SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(complement_size,complement_demand,truck_cap):
-			     2*BINS(complement_demand,truck_cap));
+			     RHS((int) complement_size, (int)complement_demand,
+				 (int) truck_cap):
+			     2*BINS((int) complement_demand, (int) truck_cap));
 	    for (k=0, cutpt = cut_list; k < num_cuts; k++,
 		    cutpt += new_cut->size)
 		  if (!memcmp(coef, cutpt, new_cut->size*sizeof(char))) break; 
@@ -642,8 +646,9 @@ int greedy_shrinking6_one(network *n, double truck_cap,
 	    new_cut->type = (set_size < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(set_size, set_demand, truck_cap):
-			     2*BINS(set_demand, truck_cap));
+			     RHS((int) set_size, (int) set_demand,
+				 (int) truck_cap):
+			     2*BINS((int) set_demand, (int) truck_cap));
 	   
 	    for (k = 0, cutpt = cut_list; k < num_cuts; k++,
 		    cutpt += new_cut->size)
@@ -680,8 +685,9 @@ int greedy_shrinking6_one(network *n, double truck_cap,
 	    new_cut->type = (complement_size  < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(complement_size,complement_demand,truck_cap):
-			     2*BINS(complement_demand,truck_cap));
+			     RHS((int) complement_size, (int)complement_demand,
+				 (int) truck_cap):
+			     2*BINS((int) complement_demand, (int) truck_cap));
 	    for (k = 0, cutpt = cut_list; k < num_cuts; k++,
 		    cutpt += new_cut->size)
 	       if (!memcmp(coef, cutpt, new_cut->size*sizeof(char))) break; 
@@ -803,8 +809,9 @@ int greedy_shrinking2_one(network *n, double truck_cap,
 	    new_cut->type = (set_size < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(set_size, set_demand, truck_cap):
-			     2*BINS(set_demand, truck_cap));
+			     RHS((int) set_size, (int) set_demand,
+				 (int) truck_cap):
+			     2*BINS((int) set_demand, (int) truck_cap));
 	    num_cuts += cg_send_cut(new_cut);
 	 }
 
@@ -830,8 +837,9 @@ int greedy_shrinking2_one(network *n, double truck_cap,
 	    new_cut->type = (complement_size  < vertnum/2 ?
 			     SUBTOUR_ELIM_SIDE:SUBTOUR_ELIM_ACROSS);
 	    new_cut->rhs =  (new_cut->type == SUBTOUR_ELIM_SIDE ?
-			     RHS(complement_size,complement_demand,truck_cap):
-			     2*BINS(complement_demand,truck_cap));
+			     RHS((int) complement_size, (int)complement_demand,
+				 (int) truck_cap):
+			     2*BINS((int) complement_demand, (int) truck_cap));
 	    
 	    num_cuts += cg_send_cut(new_cut);
 	 }
